消除编译警告

+++ main.cpp
@@ -5,7 +5,7 @@
 int main()
 {
     char key[] = "f79fa56a4cc1";
-    tinyotp::totp otp(key, strlen(key), 2);
+    tinyotp::totp otp(key, (int)strlen(key), 2);
     unsigned int code0 = otp.get_code(_time64(nullptr));
     unsigned int code1 = otp.get_code_by_step(1);
     return 0;
